+++ src/components/Breadcrumb/index.ts
@@ -1,0 +1,10 @@
+import type { App, ObjectPlugin } from 'vue';
+import BsBreadcrumb from './BsBreadcrumb';
+
+const BsBreadcrumbPlugin: ObjectPlugin = {
+    install: (app: App): void => {
+        app.component(BsBreadcrumb.name as string, BsBreadcrumb);
+    },
+};
+
+export { BsBreadcrumbPlugin, BsBreadcrumb };