+++ src/framework.ts
@@ -32,3 +32,4 @@ export * from './components/Listbox';
 export * from './components/Combobox';
 export * from './components/Modal';
 export * from './components/Notification';
+export * from './components/Breadcrumb';