+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Released: December XX, 2024
 - Reduce the number of color variants for each component on main css file
   (comply to [Bootstrap](https://getbootstrap.com/docs/5.2/getting-started/introduction/) colors with some addition)
   and put other color variants on different css file. This action greatly reduces the size of the main css file.
-- New components: **BsFontAwesome**. 
+- New components: **BsFontAwesome**, **BsBreadcrumb**. 
 - **BsIcon**, **BsIconSvg**, **BsToggleIcon**: 
   - Improve `icon` property to better accommodate **Google Material Symbols**.
   - Add new property `filled` to better accommodate **Google Material Symbols**.